streamlit_app/requirements.txt
streamlit==1.32.0
pandas==2.2.1
mysql-connector-python==8.3.0
plotly==5.20.0
